changed to standalone program using cmake to build
Started by creating a Qt Quick Application - Empty with Cmake as the
build system.  CMakeLists.txt was modified to use C++14 and the
"-fno-sized_deallocation" option was needed to run correctly.  The
generated main.cpp was only cleaned up for formatting.

The qml.qrc file lists the resources of the program, which includes all
of the .qml files needed by the program.  It also lists other resource
files that get embedded into the final program, for instance the times
up audio file needed when the timer times out.

The .gitignore was updated to the auto generated one plus an entry for
the build directory.  The previous .qmlproject file was removed.